Refactor ShortlinkController to action.
