+++ module/VuFind/config/module.config.php
@@ -225,7 +225,6 @@ $config = [
             'VuFind\Controller\RelaisController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\SearchController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\ShibbolethLogoutNotificationController' => 'VuFind\Controller\AbstractBaseFactory',
-            'VuFind\Controller\ShortlinkController' => 'VuFind\Controller\AbstractBaseWithConfigFactory',
             'VuFind\Controller\SimulatedSSOController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\SiteMapController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\SummonController' => 'VuFind\Controller\AbstractBaseFactory',
@@ -359,8 +358,6 @@ $config = [
             'search' => 'VuFind\Controller\SearchController',
             'ShibbolethLogoutNotification' => 'VuFind\Controller\ShibbolethLogoutNotificationController',
             'shibbolethlogoutnotification' => 'VuFind\Controller\ShibbolethLogoutNotificationController',
-            'Shortlink' => 'VuFind\Controller\ShortlinkController',
-            'shortlink' => 'VuFind\Controller\ShortlinkController',
             'SimulatedSSO' => 'VuFind\Controller\SimulatedSSOController',
             'simulatedsso' => 'VuFind\Controller\SimulatedSSOController',
             'SiteMap' => 'VuFind\Controller\SiteMapController',
