+++ module/VuFind/src/VuFind/Action/PluginManager.php
@@ -29,6 +29,9 @@
 
 namespace VuFind\Action;
 
+use Laminas\ServiceManager\Exception\ContainerModificationsNotAllowedException;
+use Laminas\ServiceManager\Exception\InvalidServiceException;
+use Laminas\ServiceManager\Exception\ServiceNotFoundException;
 use VuFind\ServiceManager\Factory\AutowiringFactory;
 
 use function count;
@@ -63,6 +66,7 @@ class PluginManager extends \VuFind\ServiceManager\AbstractPluginManager
      * @var array
      */
     protected $categoryAliases = [
+        'Shortlink' => 'ShortLink',
     ];
 
     /**
@@ -138,9 +142,9 @@ class PluginManager extends \VuFind\ServiceManager\AbstractPluginManager
      *
      * @return mixed
      *
-     * @throws Exception\ServiceNotFoundException If the manager does not have a service definition for the instance,
+     * @throws ServiceNotFoundException If the manager does not have a service definition for the instance,
      * and the service is not auto-invokable.
-     * @throws InvalidServiceException If the plugin created is invalid for the plugin context.
+     * @throws InvalidServiceException  If the plugin created is invalid for the plugin context.
      */
     public function get($name, ?array $options = null)
     {
